removed template Dir field & added remove empty folders from delete file (#5)

+++ src/jvmMain/kotlin/com/jaehl/codeTool/util/FileUtil.kt
@@ -9,6 +9,7 @@ import java.nio.file.StandardCopyOption
 
 interface FileUtil {
     fun createDirectory(path : Path)
+    fun renameDirectory(oldPath : Path, newPath : Path) : Boolean
     fun deleteDirectory(path : Path)
     fun createFile(path : Path)
     fun deleteFile(path : Path)
@@ -32,11 +33,24 @@ class FileUtilImp(
         }
     }
 
+    override fun renameDirectory(oldPath : Path, newPath : Path) : Boolean{
+        return try {
+            Files.move(oldPath, newPath, StandardCopyOption.REPLACE_EXISTING)
+            true
+        } catch (e : IOException) {
+            logger.error("$tag::renameDirectory ${e.message}")
+            false
+        }
+    }
+
     override fun deleteDirectory(path : Path) {
-        Files.delete(path)
+        if(Files.exists(path)) {
+            Files.delete(path)
+        }
     }
 
     override fun createFile(path: Path) {
+        walkCreateFolders(path.toFile().absolutePath.split(getPathSeparator()).dropLast(1))
         if(!Files.exists(path)) {
             Files.createFile(path)
         }
@@ -45,6 +59,7 @@ class FileUtilImp(
     override fun deleteFile(path : Path) {
         if(Files.exists(path)) {
             Files.delete(path)
+            walkDeleteEmptyFolders(path.toFile().absolutePath.split(getPathSeparator()).dropLast(1))
         } else {
             logger.error("$tag::deleteFile files does not exist \"${path.toAbsolutePath()}\"")
         }
@@ -79,9 +94,34 @@ class FileUtilImp(
         return File.separator
     }
 
+    private fun walkCreateFolders(path: List<String>, index : Int = 0){
+        if(index >= path.size) return
+        val folder = Path.of(path.subList(0, index+1).joinToString(getPathSeparator()))
+        if(!Files.exists(folder)){
+            Files.createDirectory(folder)
+        }
+        walkCreateFolders(path, index+1)
+    }
+
+    @Throws(IOException::class)
+    private fun isDirEmpty(directory: Path): Boolean {
+        Files.newDirectoryStream(directory).use { dirStream -> return !dirStream.iterator().hasNext() }
+    }
+
+    private fun walkDeleteEmptyFolders(path: List<String>, index : Int = (path.size -1)){
+        if(index < 0) return
+        val folder = Path.of(path.subList(0, index+1).joinToString(getPathSeparator()))
+        if(Files.exists(folder) && isDirEmpty(folder)){
+            deleteDirectory(folder)
+        }
+        walkDeleteEmptyFolders(path, index-1)
+    }
+
     override fun moveFile(src: Path, dest: Path) : Boolean {
         return try {
+            walkCreateFolders(dest.toFile().absolutePath.split(getPathSeparator()).dropLast(1))
             Files.move(src, dest, StandardCopyOption.REPLACE_EXISTING)
+            walkDeleteEmptyFolders(src.toFile().absolutePath.split(getPathSeparator()).dropLast(1))
             true
         } catch (e : IOException) {
             logger.error("$tag::moveFile ${e.message}")
